+++ src/parser/block/extension/fig.ts
@@ -6,6 +6,7 @@ import { segment as seg_label } from '../../inline/extension/label';
 import { segment as seg_code } from '../codeblock';
 import { segment as seg_math } from '../mathblock';
 import { segment as seg_blockquote } from '../blockquote';
+import { segment as seg_table } from './table';
 import { segment as seg_placeholder } from './placeholder';
 
 import FigParser = ExtensionParser.FigParser;
@@ -17,6 +18,7 @@ export const segment: FigParser.SegmentParser = block(validate(['[$', '$'],
       seg_code,
       seg_math,
       seg_blockquote,
+      seg_table,
       seg_placeholder,
       some(contentline),
     ]),